:sparkles: Add Interface for Edit Distance
The Edit Distance  between two strings is a measure of the similarity between the two strings. A lower edit distance indicates that the two strings are more similar, while a higher edit distance indicates that the two strings are less similar.

The interface provides a single method that calculates the similarity score in a unit provided by the user (e.g., double or integer), for two strings provided as input.