:sparkles: Add Tokenizer interface
The tokenizer interface provides a single method that provides a set of tokens generated from a single string. 

These tokens can vary based on the concrete implementation of the interface. Example tokens can include words, characters or substrings.

At the time of writing this interface only has one implementation in the library. This interface is primarily designed to be an extension point for different tokenizers that can be used with similarity calculators provided by this library.